backend-django/requirements.txt
amqp==5.1.1
appnope==0.1.3
asgiref==3.5.2
asttokens==2.0.5
backcall==0.2.0
beautifulsoup4==4.11.1
billiard==3.6.4.0
celery==5.2.6
certifi==2021.10.8
charset-normalizer==2.0.12
click==8.1.3
click-default-group==1.2.2
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
decorator==5.1.1
Django==4.0.4
django-cors-headers==3.12.0
djangorestframework==3.13.1
executing==0.8.3
greenlet==1.1.2
idna==3.3
ipython==8.3.0
jedi==0.18.1
kombu==5.2.4
matplotlib-inline==0.1.3
oauthlib==3.2.0
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
pika==1.2.1
Pillow==9.1.1
playwright==1.22.0
prompt-toolkit==3.0.29
ptyprocess==0.7.0
pure-eval==0.2.2
py-trello==0.18.0
pyee==8.1.0
Pygments==2.12.0
python-dateutil==2.8.2
pytz==2022.1
PyYAML==6.0
requests==2.27.1
requests-oauthlib==1.3.1
shot-scraper==0.14.1
six==1.16.0
soupsieve==2.3.2.post1
sqlparse==0.4.2
stack-data==0.2.0
traitlets==5.2.1.post0
trello==0.9.7.3
urllib3==1.26.9
vine==5.0.0
wcwidth==0.2.5
websockets==10.1
